feat: remove markdown-it entirely
