+++ apps/editor/test/demo/editor.html
@@ -28,7 +28,6 @@
     </div>
     <!-- Editor Dependencies -->
     <script src="../../node_modules/jquery/dist/jquery.js"></script>
-    <script src="../../node_modules/markdown-it/dist/markdown-it.js"></script>
     <script src="../../node_modules/codemirror/lib/codemirror.js"></script>
     <!-- Editor -->
     <script src="../../dist/tui-editor.js"></script>
